pyproject.toml
[project]
name = "lexai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tanish643", email = "your-atanish819@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.72.0",
    "httpx>=0.28.1",
    "pinecone>=8.1.2",
    "pydantic>=2.13.0",
    "pypdf>=6.10.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "voyageai>=0.3.7",
]

[project.scripts]
lexai = "lexai:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that hit real external APIs (Voyage, Pinecone, Gemini) — require .env keys",
]
